+++ src/components/takes/options/TakeExportMenuItem.jsx
@@ -1,9 +1,14 @@
 import { MenuItem } from '@ariakit/react';
 import { useEffect, useState } from 'react';
 
+import AddToDriveIcon from '@material-symbols/svg-400/rounded/add_to_drive.svg';
 import DownloadIcon from '@material-symbols/svg-400/rounded/download.svg';
 
-import { useTakeDownloader } from '@/serdes/UseTakeExporter';
+import { useGAPITokenHandler } from '@/lib/googleapi';
+import {
+  useTakeDownloader,
+  useTakeGoogleDriveUploader,
+} from '@/serdes/UseTakeExporter';
 import { useTakeExportedIDBKey } from '@/stores/DocumentStoreContext';
 import { getVideoBlob } from '@/stores/VideoCache';
 import MenuStyle from '@/styles/Menu.module.css';
@@ -27,8 +32,11 @@ export default function TakeCacheMenuItem({
   const menuOpen = useForceRefreshOnMenuOpen();
 
   const [cached, setCached] = useState(false);
+  const [googleDriveConnected, setGoogleDriveConnected] = useState(false);
   const idbKey = useTakeExportedIDBKey(documentId, takeId);
   const downloadTake = useTakeDownloader();
+  const uploadTake = useTakeGoogleDriveUploader();
+  const handleToken = useGAPITokenHandler();
 
   function onDownloadClick() {
     if (!idbKey) {
@@ -37,7 +45,12 @@ export default function TakeCacheMenuItem({
     downloadTake(documentId, sceneId, shotId, takeId);
   }
 
+  function onGoogleDriveClick() {
+    uploadTake(documentId, sceneId, shotId, takeId);
+  }
+
   useEffect(() => {
+    // Test video in cache
     getVideoBlob(takeId)
       .then((result) => {
         if (result) {
@@ -47,14 +60,28 @@ export default function TakeCacheMenuItem({
         }
       })
       .catch(() => setCached(false));
+
+    // Test google drive connection
+    if (!handleToken((token) => setGoogleDriveConnected(Boolean(token)))) {
+      setGoogleDriveConnected(false);
+    }
   }, [menuOpen]);
 
   return (
-    <MenuItem
-      className={MenuStyle.menuItem}
-      onClick={onDownloadClick}
-      disabled={!cached}>
-      <DownloadIcon className="w-6 h-6 fill-current" /> Download
-    </MenuItem>
+    <div className="flex flex-row">
+      <MenuItem
+        className={MenuStyle.menuItem + ' ' + 'flex-1 flex flex-row'}
+        onClick={onDownloadClick}
+        disabled={!cached}>
+        <span className="flex-1">Export to</span>
+        <DownloadIcon className="w-6 h-6 fill-current" />
+      </MenuItem>
+      <MenuItem
+        className={MenuStyle.menuItem}
+        onClick={onGoogleDriveClick}
+        disabled={!cached || !googleDriveConnected}>
+        <AddToDriveIcon className="w-6 h-6 fill-current" />
+      </MenuItem>
+    </div>
   );
 }